requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.9.14; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
django==4.1.1
idna==3.4; python_version >= '3.5'
pyyaml==6.0
requests==2.28.1; python_version >= '3.0'
sqlparse==0.4.2; python_version >= '3.5'
stripe==4.1.0
tzdata==2022.2; sys_platform == 'win32'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
